# aspnetcore/src/IO.Swagger/Controllers/DeviceApi.cs
/*
 * Sample IoT APIs
 *
 * These are public domain APIs, currently used for demo purpose.
 *
 * OpenAPI spec version: 1.0.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;

using Microsoft.AspNetCore.Authorization;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class DeviceApiController : ControllerBase
    { 
        /// <summary>
        /// 
        /// </summary>
        /// <remarks>returns all registered devices</remarks>
        /// <param name="skip">number of records to skip</param>
        /// <param name="limit">max number of records to return</param>
        /// <response code="200">All the devices</response>
        [HttpGet]
        [Route("/thirumalavenkateshc/IOTAPI/1.0.0/devices")]
        [ValidateModelState]
        [SwaggerOperation("GetDevices")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<string>), description: "All the devices")]
        public virtual IActionResult GetDevices([FromQuery]int? skip, [FromQuery]int? limit)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(List<string>));
            string exampleJson = null;
            exampleJson = "[ \"http://10.0.0.225:8080\", \"http://10.0.0.225:8080\" ]";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<List<string>>(exampleJson)
                        : default(List<string>);            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// 
        /// </summary>
        /// <param name="body"></param>
        /// <response code="200">successfully registered device</response>
        [HttpPost]
        [Route("/thirumalavenkateshc/IOTAPI/1.0.0/devices")]
        [ValidateModelState]
        [SwaggerOperation("Register")]
        public virtual IActionResult Register([FromBody]DeviceRegistrationInfo body)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200);

            throw new NotImplementedException();
        }
    }
}